requirements.txt
flask
flask-cors
python-dotenv
azure-ai-projects==1.0.0b1
azure-identity
